Add VM flag for stressing garbage collector
In the book, this is set with the preprocessor, leaving no trace if the
mode is disabled. In a real system, I would add an optimized build of
the native library that also removes all traces of debugging code while
still allowing them to be controlled at runtime instead of build time.